fix todo

+++ src/main/java/com/example/demo/components/todo/TodoRepository.java
@@ -2,5 +2,5 @@ package com.example.demo.components.todo;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface TodoRepository extends CrudRepository<Todo, Long> {
+public interface TodoRepository extends CrudRepository<Todo, Integer> {
 }
